header show info by user role and fixed some files paths
