+++ client/src/components/index.ts
@@ -1,5 +1,5 @@
 
-import LayoutMain from './Layout/LayoutMain/index'
+import LayoutMain from './layout/LayoutMain'
 
 const components = {
     LayoutMain,
