+++ client/src/components/job-cards/CardsDetail.tsx
@@ -2,23 +2,22 @@ import React from "react";
 import styled from "styled-components";
 import { Button } from "antd";
 import Title from "./Title";
-import "./";
 
-function CardsDetals3() {
+function CardsDetail() {
   return (
     <Container>
       <Title title="Full-Stack developer" />
       <TopSection>
         <div>
           <div>
-            <span>Google</span>
-            <span> -Herzeliya</span>
+            <span>checkPoint</span>
+            <span> -Tel Aviv</span>
           </div>
           <div>
             <span>active</span>
-            <span> -applicants 12</span>
+            <span> -applicants</span>
           </div>
-          <div>last date to apply : 23/4/2021</div>
+          <div>last date to apply : 23/6/2021</div>
         </div>
 
         <Button type="primary">apply job</Button>
@@ -26,23 +25,15 @@ function CardsDetals3() {
 
       <SecondSection>
         <TherdSection>
+
           <p className="jobDescription">
-            With more than a billion active listings, tens of millions of daily
-            transactions, conducted by tens of millions of sellers / hundreds of
-            millions of buyers, across more than 30 countries, and 22 different
-            currencies, comprised of a product catalog staggering in diversity -
-            eBay is a marketplace like no other that has ever existed If you’re
-            interested in joining a purpose driven community that is dedicated
-            to creating an ambitious and inclusive workplace, join eBay – a
-            company you can be proud to be a part of.
+            At DeePathology.ai, we are developing next generation tools for AI
+            in Pathology, that unleash massive AI creation, and radically
+            optimize accuracy and time utilization in pharma and diagnostics.
             <br />
-            In this role you will be a part of eBay’s global Seller Insights
-            group in support of dramatically elevating the capabilities of all
-            of eBay’s Selling. Ensuring eBay’s tens of millions of sellers have
-            access to world class, supremely intelligent tools, analytics and
-            API’s to support their needs to operate and grow. This is a key
-            pillar of eBay’s strategic vision for the company. This is a mixed
-            oriented role with 50-60% of BE vs. 40-50% frontend.
+            Our STUDIO platform lets Pathologists create AI solutions by
+            themselves, leveraging Active Learning, Interactive/Online Learning
+            and more.
           </p>
           <p>
             As a Full Stack Engineer, you will work on both the frontend and
@@ -53,17 +44,14 @@ function CardsDetals3() {
         <div>
           <h4>Minimum Qualifications:</h4>
           <ul>
-            <li>5+ years of hands-on knowledge of Java application development, testing and frameworks (Spring Boot)</li>
+            <li>Must: 3+ Years experience in Angular or Vue or React</li>
             <li>
               {" "}
               Hands-on experience with HTML5, CSS3, JavaScript, Typescript.
             </li>
-            <li>Experience with NoSQL DBs (Elasticsearch , Redis , MongoDB , etc.)</li>
             <li>Good knowledge of Python.</li>
             <li>"Let's make this happen!" attitude</li>
             <li>Team player, self learning abilities, independence.</li>
-            <li>Proven experience working in agile environment and cross functional teams</li>
-            <li>Experience with Docker/Kubernetes - Advantage</li>
           </ul>
         </div>
       </SecondSection>
@@ -71,7 +59,7 @@ function CardsDetals3() {
   );
 }
 
-export default CardsDetals3;
+export default CardsDetail;
 
 const Container = styled.div`
   height: 100vh;
