+++ client/src/components/job-cards/CardsDetails3.tsx
@@ -2,23 +2,22 @@ import React from "react";
 import styled from "styled-components";
 import { Button } from "antd";
 import Title from "./Title";
-import "./";
 
-function CardsDetail() {
+function CardsDetals3() {
   return (
     <Container>
       <Title title="Full-Stack developer" />
       <TopSection>
         <div>
           <div>
-            <span>checkPoint</span>
-            <span> -Tel Aviv</span>
+            <span>Google</span>
+            <span> -Herzeliya</span>
           </div>
           <div>
             <span>active</span>
-            <span> -applicants</span>
+            <span> -applicants 12</span>
           </div>
-          <div>last date to apply : 23/6/2021</div>
+          <div>last date to apply : 23/4/2021</div>
         </div>
 
         <Button type="primary">apply job</Button>
@@ -26,15 +25,23 @@ function CardsDetail() {
 
       <SecondSection>
         <TherdSection>
-
           <p className="jobDescription">
-            At DeePathology.ai, we are developing next generation tools for AI
-            in Pathology, that unleash massive AI creation, and radically
-            optimize accuracy and time utilization in pharma and diagnostics.
+            With more than a billion active listings, tens of millions of daily
+            transactions, conducted by tens of millions of sellers / hundreds of
+            millions of buyers, across more than 30 countries, and 22 different
+            currencies, comprised of a product catalog staggering in diversity -
+            eBay is a marketplace like no other that has ever existed If you’re
+            interested in joining a purpose driven community that is dedicated
+            to creating an ambitious and inclusive workplace, join eBay – a
+            company you can be proud to be a part of.
             <br />
-            Our STUDIO platform lets Pathologists create AI solutions by
-            themselves, leveraging Active Learning, Interactive/Online Learning
-            and more.
+            In this role you will be a part of eBay’s global Seller Insights
+            group in support of dramatically elevating the capabilities of all
+            of eBay’s Selling. Ensuring eBay’s tens of millions of sellers have
+            access to world class, supremely intelligent tools, analytics and
+            API’s to support their needs to operate and grow. This is a key
+            pillar of eBay’s strategic vision for the company. This is a mixed
+            oriented role with 50-60% of BE vs. 40-50% frontend.
           </p>
           <p>
             As a Full Stack Engineer, you will work on both the frontend and
@@ -45,14 +52,17 @@ function CardsDetail() {
         <div>
           <h4>Minimum Qualifications:</h4>
           <ul>
-            <li>Must: 3+ Years experience in Angular or Vue or React</li>
+            <li>5+ years of hands-on knowledge of Java application development, testing and frameworks (Spring Boot)</li>
             <li>
               {" "}
               Hands-on experience with HTML5, CSS3, JavaScript, Typescript.
             </li>
+            <li>Experience with NoSQL DBs (Elasticsearch , Redis , MongoDB , etc.)</li>
             <li>Good knowledge of Python.</li>
             <li>"Let's make this happen!" attitude</li>
             <li>Team player, self learning abilities, independence.</li>
+            <li>Proven experience working in agile environment and cross functional teams</li>
+            <li>Experience with Docker/Kubernetes - Advantage</li>
           </ul>
         </div>
       </SecondSection>
@@ -60,7 +70,7 @@ function CardsDetail() {
   );
 }
 
-export default CardsDetail;
+export default CardsDetals3;
 
 const Container = styled.div`
   height: 100vh;
