+++ client/src/components/job-cards/CardsDetals2.tsx
@@ -2,21 +2,20 @@ import React from "react";
 import styled from "styled-components";
 import { Button } from "antd";
 import Title from "./Title";
-import "./";
 
-function CardsDetals4() {
+function CardsDetals2() {
   return (
     <Container>
       <Title title="Full-Stack developer" />
       <TopSection>
         <div>
           <div>
-            <span>Ebay</span>
+            <span>Microsoft</span>
             <span> -Tel Aviv</span>
           </div>
           <div>
             <span>active</span>
-            <span> -applicants 41</span>
+            <span> -applicants 23</span>
           </div>
           <div>last date to apply : 23/4/2021</div>
         </div>
@@ -40,7 +39,9 @@ function CardsDetals4() {
             group in support of dramatically elevating the capabilities of all
             of eBay’s Selling. Ensuring eBay’s tens of millions of sellers have
             access to world class, supremely intelligent tools, analytics and
-            API’s to support their needs to operate and grow.
+            API’s to support their needs to operate and grow. This is a key
+            pillar of eBay’s strategic vision for the company. This is a mixed
+            oriented role with 50-60% of BE vs. 40-50% frontend.
           </p>
           <p>
             As a Full Stack Engineer, you will work on both the frontend and
@@ -51,7 +52,7 @@ function CardsDetals4() {
         <div>
           <h4>Minimum Qualifications:</h4>
           <ul>
-            <li>4+ years of hands-on knowledge of Java application development, testing and frameworks (Spring Boot)</li>
+            <li>5+ years of hands-on knowledge of Java application development, testing and frameworks (Spring Boot)</li>
             <li>
               {" "}
               Hands-on experience with HTML5, CSS3, JavaScript, Typescript.
@@ -59,12 +60,7 @@ function CardsDetals4() {
             <li>Experience with NoSQL DBs (Elasticsearch , Redis , MongoDB , etc.)</li>
             <li>Good knowledge of Python.</li>
             <li>"Let's make this happen!" attitude</li>
-            <li>BSc in computer science/software engineering or other relevant education</li>
-            <li>At least 1 year experience with NodeJS</li>
-            <li>Take ownership and help drive the team and group efforts</li>
             <li>Team player, self learning abilities, independence.</li>
-            <li>Proficiency with fundamental front end languages such as Json, XML, AJAX HTML, React, Angular, CSS and JavaScript.</li>
-            <li>Familiarity with database technology NoSql –an advantage</li>
           </ul>
         </div>
       </SecondSection>
@@ -72,7 +68,7 @@ function CardsDetals4() {
   );
 }
 
-export default CardsDetals4;
+export default CardsDetals2;
 
 const Container = styled.div`
   height: 100vh;
