+++ client/src/components/job-cards/JobCard.tsx
