+++ client/src/components/loginForm/Login.tsx
@@ -1,33 +1,38 @@
-import React from 'react'
-import { Form, Input, Button} from 'antd';
+import React,{useState} from 'react'
+import { Form, Input, Button, Select } from 'antd';
 import { connect } from 'react-redux';
 import usersActions from '../../redux/actions/user.actions';
 import LogIn from '../../utils/login.utils';
+const { Option } = Select;
 
-const { loginUser,setTokenLocal } = LogIn
+const { loginUser, setTokenLocal } = LogIn
 const { setUserData } = usersActions.usersActions;
-const mapDispatchToProps = (dispatch: any) => ({
-  setUserData: (data: Object) => { dispatch(setUserData(data)) }})
-const mapStateToProps = (state: any) => {return {userData: state.user.userData}}
+
+const mapDispatchToProps = (dispatch: any) => ({setUserData: (data: Object) => { dispatch(setUserData(data)) }})
+const mapStateToProps = (state: any) => { return { userData: state.user.userData } }
+
 const Login = (): any => {
+    const [role, setRole] = useState('student')
     const layout = { labelCol: { span: 8 }, wrapperCol: { span: 16 } }
     const tailLayout = { wrapperCol: { offset: 8, span: 16 } }
 
-    const onFinish = async (values: any) => { 
-        const email:any =  values.email
-        const password:any =  values.password
-        const resFromLogin = await loginUser({email,password},'student')
-        if(resFromLogin.success){
+    const onFinish = async (values: any) => {
+        const email: any = values.email
+        const password: any = values.password
+        const resFromLogin = await loginUser({ email, password }, role)
+        if (resFromLogin.success) {
             const token = resFromLogin.token
             setTokenLocal(token)
             setUserData(token)
             window.location.reload();
 
-        }else{alert('failed!')}
-        
+        } else { alert('failed!') }
+
     }
     const onFinishFailed = (errorInfo: any) => { console.log('Failed:', errorInfo) }
-
+    const handleChange = (value: any) =>{
+        setRole(value)
+    }
     return (
         <Form
             {...layout}
@@ -53,9 +58,16 @@ const Login = (): any => {
             </Form.Item>
 
 
-            <Form.Item  {...tailLayout} >
-                <Button type="primary" htmlType="submit"> התחבר </Button>
-            </Form.Item>
+            <div style={{ display: 'flex', justifyContent: 'space-between', }}>
+                <Form.Item  {...tailLayout} >
+                    <Button type="primary" htmlType="submit"> התחבר </Button>
+                </Form.Item>
+                <Select defaultValue="student" style={{ width: 120, height: '110%' }} onChange={handleChange}>
+                    <Option value="student">Student</Option>
+                    <Option value="hr">Hr</Option>
+                    <Option value="admin">Admin</Option>
+                </Select>
+            </div>
         </Form>
     );
 };
