+++ server/api/register/register_ctrl.js
@@ -258,6 +258,7 @@ async function loginUser(req, res) {
     const role = req.params.Role && req.params.Role.toLowerCase()
     const { email, password } = user
     const query = { email }
+    console.log(query);
     if (!role) return res.status(404).json({
         success: false,
         message: 'role is required'
@@ -272,7 +273,6 @@ async function loginUser(req, res) {
     const getDocSuccessCb = async(data) => {
         const { _id, name, email } = data;
         const passwordFromDB = data.password;
-        console.log(data);
         if (!data.isAuth) return res.status(400).json({
             success: false,
             message: 'unsigned user on loginUser'
